refact(BuyingPrice): 코드의 통일성을 위해 validator를 여기에 추가
