--- src/main/java/lotto/util/Validator.java
@@ -1,13 +1,0 @@
-package lotto.util;
-
-
-public abstract class Validator<T> {
-
-    private static String ERROR_HEADER = "[ERROR] ";
-
-    protected static void throwError(String errorMessage) {
-        throw new IllegalArgumentException(ERROR_HEADER + errorMessage);
-    }
-
-    abstract void validate(T param);
-}
